Move Calloc and Free to y
