+++ y/calloc.go
@@ -2,7 +2,7 @@
 // of this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 
-package manual
+package y
 
 // #include <stdlib.h>
 import "C"
