+++ y/calloc_32bit.go
@@ -4,7 +4,7 @@
 
 // +build 386 amd64p32 arm armbe  mips mipsle mips64p32 mips64p32le ppc sparc
 
-package manual
+package y
 
 const (
 	// MaxArrayLen is a safe maximum length for slices on this architecture.
