+++ y/calloc_64bit.go
@@ -4,7 +4,7 @@
 
 // +build amd64 arm64 arm64be ppc64 ppc64le mips64 mips64le s390x sparc64
 
-package manual
+package y
 
 const (
 	// MaxArrayLen is a safe maximum length for slices on this architecture.
